refactor: Gear Closet -> move logic/mutations to parent component
